Show links to PG eBooks in sidebar
This is the initial chunk of code to allow displaying PG eBook links
for OpenLibrary records that have Project Gutenberg numbers associated
with them. If multiple PG numbers are associated with a given record,
the largest PG number is used for the links, with the assumption that
more recent PG eBooks will have better formatting than older ones.

This code relies on the Gutenberg python module that provides fast
access to PG URLs after building a local cache of the full PG catalog.
Additions were made to the Gutenberg project to allow fetching the
URLs and more fine-grained control over the local cache. These additions
are complete but as of this commit not yet accepted into the master repo.